+++ packages/google_maps_flutter/google_maps_flutter_ios_sdk10/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.18.7
+
+* Converts circle, polygon, and polyline controllers to Swift.
+
+## 2.18.6
+
+* Converts parts of the implementation to Swift.
+
 ## 2.18.5
 
 * Fixes a potential compilation issue in tile downscaling.